Document minimum resources for Sdext operation
The baseline strategy for accessing capabilities in integer registers
requires 4 instructions in the program buffer, 1 XLEN's worth of data
register shadows, and 1 dscratch register, all of which are optional
in the debug spec.  Document that Zcheri needs them.

Forbid hartinfo.dataaccess=1 since it depends on loads and stores using
x0 as a base.