Deploy GitHub pages on every commit to main
This adds a note to the generated HTML that this page is a snapshot and the
PDF should be treated as the latest source. Automatically deploying each
commit to https://riscv.github.io/riscv-cheri/ makes it easier to read
the latest snapshot version of the spec without requiring a new release.

Fixes: https://github.com/riscv/riscv-cheri/issues/74