Add IntelliJ editing plugin support (#120)
Move attribute definitions to a separate .adoc file, so that they
can be read by the IntelliJ plugin. 
This makes it easier for me to edit the files and will not have
any functional changes. Opening this as a pull request to ensure
I don't conflict with anyone currently editing the asciidoc attributes.