Share CSR addresses between foo and fooc
Extending the ISA width by adding new names for every operation and
register is a MIPSism.  It is far more consistent with the RISC-V base
ISA to reuse the same instructions with a contextually defined width.
This also saves a not entirely negligible fraction of the standard CSR
encoding space.

It was already the case that only one size of CSR can be used in any
given mode, so this does not reduce software flexibility.

Signed-off-by: Stefan O'Rear <sorear@fastmail.com>