Use resumereq instead of DRET
DRET is not a standard instruction.  Its only mention in the debug spec
is in a non-normative section describing a single possible
implementation, in a clause describing how the debug module might
internally implement the dmstatus.resumereq field.  While DRET has a
reserved opcode for implementations which use it or something like it,
it is itself a custom extension whenever it appears.

Describe CHERI functions in terms of the architectural
dmstatus.resumereq instead of the non-architectural DRET.

Signed-off-by: Stefan O'Rear <sorear@fastmail.com>