Redefine MODESW using loose association
We now have three models for the implementation of MODESW: storing the
mode in the pcc metadata, as ISAv9, storing the mode in the pcc LSB, as
Morello, and storing the mode in a loosely associated bit.

Using the metadata is robust but uses a scarce metadata bit (especially
on RV32) for a bit with extremely low entropy, since legacy code by
definition does not make heavy use of code capabilities.

Storing the mode in the address LSB is superficially attractive, since
it is unused for code capabilities, but it is incompatible with existing
RISC-V design decisions (silently masking LSB on legacy mode jalr and
writes to xepc).

Were it made possible to use the LSB of function pointers and return
addresses in that way, it would be most useful for mode switches which
do not affect the ABI and can be mixed freely at the level of possibly
indirect calls. We have two very different ABIs and only one useful
instruction set for each ABI, so there is little to be gained by
decoupling the ABI from the implementation instruction set of a
function.

Assuming, then, that the entry instruction set of a function is
determined by the ABI used to call it, it is natural to decouple the
choice of instruction set from the function address. We define in this
proposal that code which is jumped to via a capability always uses
Capability mode; to jump to code in Legacy mode, use a trap return or a
trampoline within the legacy pcc bounds.

By requiring trap handlers to use Capability mode and not storing mode
metadata for CSRs other than xepc, this proposal achieves the smallest
architectural state of the three options, only 4 bits for D+M+S+U.

Fixes #31.
Fixes #34.
Fixes #63.

Signed-off-by: Stefan O'Rear <sorear@fastmail.com>