SBI binding for CHERI
This is considered to be in scope on the theory that this repository
already contains patches to various specifications that touch CHERI, and
the SBI is an expected part of the supervisor execution environment with
the same status as ISA specifications.

Permission checks for SBI ecalls are required to support
compartmentalization within S-mode. Using ASR, specifically, is a
nontrivial design decision with rationale laid out in
https://lists.riscv.org/g/sig-cheri/message/68 . Generating a CHERI
exception instead of returning a SBI error code makes the behavior
consistent with ISA permission checks, improves error location for
debugging, and defers committing to purely software permission checks.

Using capabilities for hart start and resume avoids requiring the SBI to
be able to mint Infinity capabilities at runtime and is expected to
simplifiy the implementation.

There is an open question whether Zcheri_legacy can be implemented in a
way which maintains simultaneous upward compatibility with
Zcheri_purecap and the non-CHERI base ISA. The chosen definitions here
are not expected to be particularly useful, but avoid introducing new
problems.

Signed-off-by: Stefan O'Rear <sorear@fastmail.com>