Drop separate almighty capability encoding
This was intended to allow a single root once we have sealing
capabilities. However, it makes the encoding a bit awkward and we don't
have the same constraint with future sealing caps that they need to be
derived from a single root as we do for RW+RX capabilities. Those two
roots can still be subset from the Almighty ASR capability.