Remove JALR.MODE (#167)
This is split out from #130, I only made a minor change to the pull
request description:

In Capability mode this can be expressed by a three-instruction sequence
of auipc, scaddr, and jalr. While this is not precisely the same (the
link register is a capability not an address) it behaves almost exactly
the same.
Purecap ABIs are expected to use capabilities, possibly sealed, as
function pointers, and relative vtables and PIC switch tables do not
benefit from jalr.mode over an instruction sequence using cadd only.

In Legacy mode with Zcheri_mode, switch to Capability mode before
adjusting pcc metadata and switch back afterward. If the legacy code
could have been exited without a trap, capability-aware code within the
same pcc region must already have existed, otherwise the trap return can
be used for the mode switch.

It is no longer possible without Zcheri_mode to transfer control to a
sentry in Legacy mode.
However, #161 merges Zcheri_mode into Zcheri_legacy, so this is no
longer a concern.

Signed-off-by: Stefan O'Rear <sorear@fastmail.com>
Co-authored-by: Stefan O'Rear <sorear@fastmail.com>