Clarify the ACPERM note on new permissions
This note should refer to newly added permissions. Whether it is a
subset or superset doesn't really matter, the problem we are trying to
deal with here is that software that is not aware of the new permission
expects the clearing to behave as before.

New wording suggested by @LawrenceEsswood