Clarify CHERI page fault dependence on PTE.U
It is possible to be in machine or supervisor mode and still read pages with PTE.U set, because of mstatus.MPRV and mstatus.SUM.

The table makes clear the condition is PTE.U not the current privilege, so this updates the text to match it.

I also removed "optionally" from the second scheme because the optionality is why there are two schemes in the first place. I removed the footnote link from there too because that footnote isn't specific to the second scheme. Arguably it shouldn't be a footnote at all.