Proofreading and restructuring some sections for a better logical flow (#622)
The intention of this PR is to move towards talking about the PCC and
data handling with a better separation as requested by Nick Kossifidis.

At the moments it's just typographical improvements, and a bit of
restructuring (e.g. all the text about valid tags is in one section now
instead of split between two).

---------

Signed-off-by: Alexander Richardson <mail@alexrichardson.me>
Co-authored-by: Alexander Richardson <mail@alexrichardson.me>