Improve description of EF=0 and EF=1 cases for clarity. (#641)
The phrase "for regions less than 2^MW-2" wasn't true for MXLEN=32, but
also was not clear. EF=1 was to be used only in exactly the cases where
length < 2^MW-2 for MXLEN=64 or 2^MW-1 for MXLEN=32, which wasn't quite
what it was saying. On closer inspection, a number of other statements
here were imprecise; hopefully this is an improvement.