remove exception checks from branches and jumps (#644)
https://github.com/riscv/riscv-cheri/issues/636

I was on the fence about whether we should make this change, so made the
PR to explore the implications.
TBH it's quite a nice simplification, so I think  we should go for it.

If we find a specific security hole during compartment switching then we
can always add a control bit somewhere to control exception behaviour
for that specific case, without requiring exceptions on _every_ branch
and jump.

---------

Signed-off-by: Tariq Kurd <tariqandlaura@gmail.com>
Co-authored-by: Alexander Richardson <mail@alexrichardson.me>