Use "capability tag" instead of "valid tag"
ARC feedback was:

Using "Valid" as the name for the tag is problematic.
The register contents are still usable even when this bit is clear.
Also, valid is used in the name but also to explain the state.
OK to use "valid capability" , shouldn't use "valid tag".
Suggest "capability tag" as full name, and just "ytag" or "ctag" as
short form.