Address latest round of unpriv ARC feedback (#675)
The biggest change here is that we no use "capability tag"
instead of "valid tag" since the ARC feedback was:

> Using "Valid" as the name for the tag is problematic.
The register contents are still usable even when this bit is clear.
Also, valid is used in the name but also to explain the state.
OK to use "valid capability" , shouldn't use "valid tag".
Suggest "capability tag" as full name, and just "ytag" or "ctag" as
short form.

I've also removed any mentions of using *x* and *c* registers.
We don't access different registers, the only difference is whether we
use only the XLEN subset or the full register.

Co-authored-by: Tariq Kurd <tariq.kurd@codasip.com>