Use y-prefixed general purpose register names instead of c-prefix
This was requested by ARC and one benefit here is that we no longer have
the ambiguity in the c.jalr expansion: `jalr y1, 0(cs1)` makes it
obvious that the first register argument is a fixed `ra` value and the
latter is the `cs1` variable from the encoding.

Use attributes so we can easily change this depending on ARC feedback.

Reviewed By: tariqkurd-repo

Pull Request: https://github.com/riscv/riscv-cheri/pull/687