Refactor CLRPERM rules (#742)
Move common rules out into the architectural specification, not the encoding,
and clarify that extensions and encodings may each add new rules.  Adjust
cross-references to follow along.

For the appendix's RV32 encoding, rework the procedural transform (necessary
because of the negative logic used, under which free-order fixed-pointing
rewriting is not a confluent strategy).  There's a script, misc/rv32perms.hs,
that verifies that the defined compressed permission sets and the given
procedure obey a number of essential properties.