Clarify exactly which extensions are stable and what we're ratifying for v1.0 (#808)
Make the ratification package clearly defined.
Put extensions into v1.0 ratification, or future categories.
All extensions are in the standalone spec.
Only v1.0 ratification extensions are in the unpriv and priv specs.

As a result, CHERIoT is included in the standalone spec only.

(all content is in the unpriv/priv specs on an ifdef)

---------

Signed-off-by: Tariq Kurd <tariq.kurd@codasip.com>
Co-authored-by: Nathaniel Wesley Filardo <VP331RHQ115POU58JFRLKB7OPA0L18E3@cmx.ietfng.org>
Co-authored-by: Alex Richardson <alexrichardson@google.com>