Zylevels1: remove optional imposition on YPERMC
YPERMC is presently phrased as requiring a check that the AP and SDP fields are
not changed under seal, but Zylevels1 introduces a GLobal flag that is
permission-like in its monotonicity but is not "under seal".

We had considered the possibility of having YPERMC require that, if its input
capability operand is sealed, then its scalar operand may at most request
clearing GLobal (if Zylevels1 is implemented) or must be zero (if not).  But
this is somewhat non-compositional and perhaps confusing, so just remove the
text in question.