zyhybrid: clarify role of rs1 in (Non-CHERI) Address Mode
As soon as DDC is used as the authorizing capability, the metadata,
incl. the tag bit, of rs1 becomes irrelevant. Only the address is
sourced from rs1.

While this behavior seems natural, and the specification certainly
gives that impression, I had to convince myself by looking at CHERI's
SAIL model [1] that my interpretation of the specification is
correct. Therefore, it seems sensible to spell it out once in prose.

1: https://github.com/CTSRD-CHERI/sail-cheri-riscv/blob/bb07488dac8322fc4a653a01d049e16fa05c9065/src/cheri_addr_checks.sail#L182-L189

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>